+++ packages/client/CHANGELOG.md
@@ -24,6 +24,7 @@
 - BREAKING: `getBody` and `getHeader` only accept block-hashes.
 - BREAKING: renamed `watchBlockBody` to `getBlockBody$`
 - BREAKING: `watchValue` options are now consistent with `watchEntries` options, and also with the rest of query options.
+- BREAKING: rename chainspecs `ksmcc3` => `kusama`, `westend2` => `westend`
 
 - **WS-Provider:**
   - BREAKING: This export has been moved in favor of `polkadot-api/ws`
@@ -41,6 +42,7 @@
 
 - BREAKING: Removed `RuntimeToken` and `CompatibilityToken`
 - BREAKING: polkadot-sdk-compat
+- BREAKING: Rococo chainspecs
 
 - **polkadot-api/utils**:
   - `jsonPrint` export